backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
gunicorn==21.2.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.3
python-dotenv==1.0.0
openpyxl==3.1.2
httpx==0.26.0
cryptography==41.0.7
reportlab==4.4.7
xlsxwriter==3.2.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
